+++ mutraction-dom/src/choose.ts
@@ -13,18 +13,16 @@ function getEmptyText() {
 }
 
 export function choose(...choices: ConditionalElement[]): Node {
-    let current: ChildNode = getEmptyText(); 
-    let currentNodeGetter: () => ChildNode = getEmptyText;   
-    
+    let current: ChildNode | undefined;
+    let currentNodeGetter: () => ChildNode;   
+
     // Flag is set when condition value caused the resolved node to change.
     // The effect needs to be disposed when the node is cleaned up, but only without the flag.
     let conditionChanging = false;
     function dispose() {
         if (!conditionChanging) sub.dispose();
     }
-    // this effect subscription probably doesn't need to be disposed
-    // since mu:if/mu:else can only be applied to HTMLElement, which has its own cleanup
-    // in other words, we expect that nodeGetter() always directly calls element()
+
     const sub = effect(function chooseEffect() {
         let newNodeGetter: (() => ChildNode) | undefined;
         for (const { nodeGetter, conditionGetter } of choices) {
@@ -39,18 +37,21 @@ export function choose(...choices: ConditionalElement[]): Node {
         // e.g. `model.x > 1` stays true when model.x changes from 2 to 3
         // but we don't want to rebuild the node
         if (newNodeGetter !== currentNodeGetter) {
-            conditionChanging = true;
-            cleanup(current);
-            conditionChanging = false;
+            if (current) {
+                conditionChanging = true;
+                cleanup(current);
+                conditionChanging = false;
+            }
 
             currentNodeGetter = newNodeGetter;
             const newNode = currentNodeGetter();
-            current.replaceWith(newNode);
+            current?.replaceWith(newNode);
 
             registerCleanup(newNode, { dispose });
             current = newNode;
         }
     }, suppress);
 
+    if (!current) throw Error("Logical error in choose() for mu:if.  No element assigned after first effect invocation.");
     return current;
 }