+++ mutraction-dom/src/index.ts
@@ -7,7 +7,6 @@ export { Swapper } from './swapper.js';
 export { Tracker, TrackerOptions, defaultTracker, track } from './tracker.js';
 export { PropReference, createOrRetrievePropRef } from './propref.js';
 export { effect } from './effect.js';
-export { DependencyList } from './dependency.js'
 export { Router } from './router.js';
 export { makeLocalStyle } from './makeLocalStyle.js';
 export { untrackedClone } from './untrackedClone.js';